PAE-1538: extract validation-failures view model from controller
